pyproject.toml
[tool.poetry]
name = "trade-framework"
version = "0.1.0"
description = ""
authors = ["otavius <developer.otavius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
